findbugs: fix 'default encoding' warnings
Found reliance on default encoding in
  nl.mindef.c2sc.nbs.olsr.pud.uplink.server.dao.impl.ClusterLeaderMsgsImpl.print(OutputStream): String.getBytes()

Found a call to a method which will perform a byte to String (or String
to byte) conversion, and will assume that the default platform encoding
is suitable. This will cause the application behaviour to vary between
platforms. Use an alternative API and specify a charset name or Charset
object explicitly.

Signed-off-by: Ferry Huberts <f.huberts@mindef.nl>