findbugs: make some inner classes static
Should
  nl.mindef.c2sc.nbs.olsr.pud.uplink.server.dao.expiry.ExpireNodesTimer$ExpiryTimerTask
be a _static_ inner class?

This class is an inner class, but does not use its embedded reference to
the object which created it.  This reference makes the instances of the
class larger, and may keep the reference to the creator object alive
longer than necessary.  If possible, the class should be made static.

Signed-off-by: Ferry Huberts <f.huberts@mindef.nl>